Fix model and GCN compatability

+++ Misc/drop_features.py
@@ -12,9 +12,9 @@ class DropFeatures(BaseTransform):
         self.one_hot_encoder = OneHotDegree(emb_dim - 1)
 
     def __call__(self, data: Data):
-        data.x = torch.zeros([data.x.shape[0], 0])
+        data.x = torch.zeros([data.x.shape[0], 0], dtype=torch.float32)
         data = self.one_hot_encoder(data)
-        data.edge_attr = torch.zeros([data.edge_attr.shape[0], self.emb_dim])
+        data.edge_attr = torch.zeros([data.edge_attr.shape[0], self.emb_dim], dtype=torch.float32)
         return data
     
     def __repr__(self) -> str:
